requirements.txt
fastapi[standard]
openai
google-generativeai
pydantic
python-dotenv
supabase
flask
chainlit